Multiple Linear Regression
We try out the application of multiple linear regression using data about an organisation which shows the amount of profit generated in each of the three states where they set up business. Their budgets are outlined under three main departments, namely:
- Research
- Operations &
- Marketing
We attempt to train using supervised learning approach, specifically Multiple Linear Regression. We go an extra mile to optimise out dependent variables by a method of Backward elimination so as to achieve the closes level of prediction in the end.

Two IDEs were used to test the level of interaction: Google Colab and Spyder. Their outputs including an excel copy of the data are shared here.